# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion RH API — ARTF",
        "description": "API REST de gestion RH (intégration administrative des agents).\n\n## Authentification\n\n1. Appeler `POST /api/login` avec email / mot de passe\n2. Copier le `token` retourné\n3. Cliquer **Authorize** en haut à droite\n4. Saisir : `Bearer {token}` (ou uniquement le token selon le client)\n\nComptes de test (seeders) :\n- `admin@arft.cg` / `Admin@2026`\n- `rh@arft.cg` / `Rh@2026`\n\n## Convention de réponse\n\n```json\n{ \"data\": {}, \"message\": \"...\" }\n```\n\nErreurs : `{ \"success\": false, \"message\": \"...\", \"errors\": {} }`",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API (variable L5_SWAGGER_CONST_HOST)"
        }
    ],
    "paths": {
...
        "/api/integration/agents": {
            "get": {
                "tags": [
                    "Intégration — Agents"
                ],
                "summary": "Liste des agents",
                "operationId": "listAgents",
                "responses": {
                    "200": {
                        "description": "Liste",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgentListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Intégration — Agents"
                ],
                "summary": "Créer un agent (+ dossier d'intégration auto)",
                "description": "Crée la fiche agent et initialise automatiquement un dossier d'intégration.",
                "operationId": "storeAgent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgentCreateResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error422"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/integration/agents/{id}": {
            "get": {
                "tags": [
                    "Intégration — Agents"
                ],
                "summary": "Détail d'un agent",
                "operationId": "showAgent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error404"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Intégration — Agents"
                ],
                "summary": "Mettre à jour un agent",
                "operationId": "updateAgent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error422"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/integration/agents/{id}/matricule": {
            "patch": {
                "tags": [
                    "Intégration — Agents"
                ],
                "summary": "Modifier le matricule d'un agent",
                "operationId": "modifierMatriculeAgent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignerMatriculeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error422"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion",
                "description": "Retourne un token Sanctum. Utiliser ensuite Authorize → Bearer {token}.",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error422"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion",
                "description": "Révoque le token Sanctum courant.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Profil utilisateur connecté",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Profil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@arft.cg"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "Admin@2026"
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Administrateur"
                            },
                            "email": {
                                "type": "string",
                                "example": "admin@arft.cg"
                            },
                            "agent_id": {
                                "type": "integer",
                                "example": null,
                                "nullable": true
                            },
                            "is_active": {
                                "type": "boolean",
                                "example": true
                            },
                            "token": {
                                "type": "string",
                                "example": "1|xxxxxxxxxxxxxxxx"
                            }
                        },
                        "type": "object"
                    },
                    "message": {
                        "type": "string",
                        "example": "Connexion réussie"
                    }
                },
                "type": "object"
            },
...
            "UserResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Error401": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Non authentifié. Veuillez vous connecter."
                    }
                },
                "type": "object"
            },
...
            "Error404": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Ressource introuvable."
                    }
                },
                "type": "object"
            },
            "Error422": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Les données fournies sont invalides."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "Le champ email est obligatoire."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "MessageResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    }
                },
                "type": "object"
            },
...
            "AgentCreateResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "agent": {
                                "$ref": "#/components/schemas/Agent"
                            },
                            "dossier": {
                                "$ref": "#/components/schemas/DossierIntegration"
                            }
                        },
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AgentListResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Agent"
                        }
                    }
                },
                "type": "object"
            },
            "AgentRequest": {
                "required": [
                    "nom",
                    "prenom",
                    "date_naissance",
                    "genre",
                    "type_integration_id"
                ],
                "properties": {
                    "nom": {
                        "type": "string",
                        "example": "MOUKOKO"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Jean"
                    },
                    "date_naissance": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-12"
                    },
                    "lieu_naissance": {
                        "type": "string",
                        "example": "Brazzaville",
                        "nullable": true
                    },
                    "nationalite": {
                        "type": "string",
                        "example": "Congolaise",
                        "nullable": true
                    },
                    "genre": {
                        "type": "string",
                        "example": "M",
                        "enum": [
                            "M",
                            "F"
                        ]
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+242066123456",
                        "nullable": true
                    },
                    "email_personnel": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "numero_cnss": {
                        "type": "string",
                        "nullable": true
                    },
                    "rib_bancaire": {
                        "type": "string",
                        "nullable": true
                    },
                    "diplome_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "grade_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "categorie_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "echelon_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "fonction_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "type_integration_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "AgentResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Agent"
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AssignerMatriculeRequest": {
                "required": [
                    "matricule"
                ],
                "properties": {
                    "matricule": {
                        "type": "string",
                        "example": "ARTF-2026-001"
                    }
                },
                "type": "object"
            },
...
            "DossierIntegration": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "reference": {
                        "type": "string",
                        "example": "INT-2026-0001"
                    },
                    "statut": {
                        "type": "string",
                        "example": "BROUILLON"
                    },
                    "statut_label": {
                        "type": "string",
                        "example": "Brouillon"
                    },
                    "date_demande": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-08-10",
                        "nullable": true
                    },
                    "poste_demande": {
                        "type": "string",
                        "example": "Chargé de mission",
                        "nullable": true
                    },
                    "nombre_postes": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "motif": {
                        "type": "string",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "type_integration_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "demandeur_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "agent_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Sanctum obtenu via POST /api/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}